App - Remove `Aura` + Add `DataFileManager`
