+++ src/app/datafilebase.cpp
@@ -14,7 +14,7 @@ namespace Nickvision::App
         {
             throw std::invalid_argument("Key must not be empty.");
         }
-        m_path = UserDirectories::get(UserDirectory::ApplicationConfig, appName) / (key + ".json");
+        m_path = UserDirectories::get(UserDirectory::ApplicationConfig, appName) / (m_key + ".json");
         if (std::filesystem::exists(m_path))
         {
             std::ifstream in{ m_path };
