+++ src/app/datafilemanager.cpp
@@ -1,0 +1,10 @@
+#include "app/datafilemanager.h"
+
+namespace Nickvision::App
+{
+    DataFileManager::DataFileManager(const std::string& appName)
+        : m_appName{ appName }
+    {
+        
+    }
+}
